Limit the checkCORSSupport to older versions of Android
Long polling is not the optimal way of streaming on android, especially not on newer versions of android.
This commit is to limit the legacy behaviour of using long polling for android browsers to older versions of android.
If the version major version is found to be 0, 1, or 2 we use the old behaviour.
If were unable to parse the version it's assumed to be a new version.