Support Websockets blocked by CSP
If the application supports the Content Security Policy, it is possible - depending on the configuration - that all Websockets or Websockets for specific hosts are blocked on client side. Until now, Atmosphere just raised an error in this case, but it did not switch to the fallback transport.
This PR extends the onerror method of the Websocket logic by a reconnect to the fallback transport, e.g. long-polling.